refactor(crafting): extract shared QualitySlider component
